rename types to defs
