+++ src/autoPVS1.py
@@ -2,10 +2,10 @@
 
 import typer
 
+from src.defs.autopvs1 import PVS1Prediction, PVS1PredictionSeqVarPath
 from src.genome_builds import GenomeRelease
 from src.seqvar import SeqVar, SeqVarResolver
 from src.seqvar_pvs1 import SeqVarPVS1
-from src.types.autopvs1 import PVS1Prediction, PVS1PredictionSeqVarPath
 
 
 class AutoPVS1:
