+++ tests/api/test_annonars.py
@@ -1,9 +1,9 @@
 import responses
 
 from src.api.annonars import AnnonarsClient
+from src.defs.annonars import AnnonarsRangeResponse
 from src.genome_builds import GenomeRelease
 from src.seqvar import SeqVar
-from src.types.annonars import AnnonarsRangeResponse
 from tests.utils import get_json_object
 
 #: Example sequence variant
