+++ tests/api/test_dotty.py
@@ -1,8 +1,8 @@
 import responses
 
 from src.api.dotty import DottyClient
+from src.defs.dotty import DottySpdiResponse
 from src.genome_builds import GenomeRelease
-from src.types.dotty import DottySpdiResponse
 from tests.utils import get_json_object
